+++ src/main/java/at/porscheinformatik/seleniumcomponents/SeleniumMatcher.java
@@ -4,20 +4,18 @@ package at.porscheinformatik.seleniumcomponents;
  * A matcher for {@link SeleniumAsserts}.
  *
  * @author ham
- * @param <T> the type of the item
+ * @param <Any> the type of the item
  */
-public interface SeleniumMatcher<T>
+public interface SeleniumMatcher<Any>
 {
 
     /**
-     * Evaluates the matcher for argument <var>item</var>. <br>
-     * This method matches against Object, instead of the generic type T. This is because the caller of the Matcher does
-     * not know at runtime what the type is (because of type erasure with Java generics). It is down to the
-     * implementations to check the correct type.
+     * Evaluates the matcher for argument <var>item</var>. This matcher has the same erasure type as the Hamcrest
+     * matcher.
      *
      * @param item the object against which the matcher is evaluated.
      * @return <code>true</code> if <var>item</var> matches, otherwise <code>false</code>.
      */
-    boolean matches(T item);
+    boolean matches(Any item);
 
 }