+++ src/main/java/at/porscheinformatik/seleniumcomponents/component/AbstractTableCellComponent.java
@@ -1,10 +1,10 @@
 package at.porscheinformatik.seleniumcomponents.component;
 
 import at.porscheinformatik.seleniumcomponents.AbstractSeleniumComponent;
-import at.porscheinformatik.seleniumcomponents.SeleniumActions;
 import at.porscheinformatik.seleniumcomponents.SeleniumComponent;
 import at.porscheinformatik.seleniumcomponents.SeleniumException;
 import at.porscheinformatik.seleniumcomponents.SeleniumUtils;
+import at.porscheinformatik.seleniumcomponents.Utils;
 import at.porscheinformatik.seleniumcomponents.WebElementSelector;
 
 /**
@@ -24,7 +24,7 @@ public abstract class AbstractTableCellComponent extends AbstractSeleniumCompone
     {
         String colspan = getAttribute("colspan");
 
-        if (SeleniumUtils.isEmpty(colspan))
+        if (Utils.isEmpty(colspan))
         {
             return 1;
         }
@@ -41,6 +41,6 @@ public abstract class AbstractTableCellComponent extends AbstractSeleniumCompone
 
     public String getText()
     {
-        return SeleniumActions.getText(this);
+        return SeleniumUtils.getText(this);
     }
 }